Move a util function from define.rs to util
